feat: cleanup c-kzg kzgsetting
