+++ crates/primitives/Cargo.toml
@@ -32,8 +32,9 @@ bitvec = { version = "1", default-features = false, features = ["alloc"] }
 bitflags = { version = "2.6.0", default-features = false }
 
 # For setting the CfgEnv KZGSettings. Enabled by c-kzg flag.
-c-kzg = { version = "2.0.0", default-features = false, optional = true }
-once_cell = { version = "1.19", default-features = false, optional = true }
+c-kzg = { version = "1.0.3", default-features = false, optional = true, features = [
+    "ethereum_kzg_settings",
+] }
 
 # Optionally use `kzg-rs` for a pure Rust implementation of KZG.
 kzg-rs = { version = "0.1", default-features = false, features = [
@@ -42,7 +43,6 @@ kzg-rs = { version = "0.1", default-features = false, features = [
 
 # utility
 enumn = "0.1"
-derive_more = { version = "0.99", optional = true }
 cfg-if = "1"
 dyn-clone = "1.0"
 
@@ -110,6 +110,7 @@ optional_beneficiary_reward = []
 rand = ["alloy-primitives/rand"]
 
 # See comments in `revm-precompile`
-c-kzg = ["dep:c-kzg", "dep:once_cell", "dep:derive_more"]
-# `kzg-rs` is not audited but useful for `no_std` environment, use it with causing and default to `c-kzg` if possible.
-kzg-rs = ["dep:kzg-rs", "dep:once_cell", "dep:derive_more"]
+c-kzg = ["dep:c-kzg"]
+# `kzg-rs` is not audited but useful for `no_std` environment.
+#   use it with causing and default to `c-kzg` if possible!
+kzg-rs = ["dep:kzg-rs"]
