+++ crates/primitives/src/kzg.rs
@@ -1,16 +1,31 @@
-mod env_settings;
-mod trusted_setup_points;
-
 cfg_if::cfg_if! {
     if #[cfg(feature = "c-kzg")] {
         pub use c_kzg::KzgSettings;
+        /// KZG Settings that allow us to specify a custom trusted setup.
+        /// or use hardcoded default settings.
+        #[derive(Debug, Clone, Default, PartialEq, Eq )]
+        pub enum EnvKzgSettings {
+            /// Default mainnet trusted setup
+            #[default]
+            Default,
+            /// Custom trusted setup.
+            Custom(std::sync::Arc<c_kzg::KzgSettings>),
+        }
+
+        impl EnvKzgSettings {
+            /// Return set KZG settings.
+            ///
+            /// In will initialize the default settings if it is not already loaded.
+            pub fn get(&self) -> &c_kzg::KzgSettings {
+                match self {
+                    Self::Default => {
+                        c_kzg::ethereum_kzg_settings()
+                    }
+                    Self::Custom(settings) => settings,
+                }
+            }
+        }
     } else if #[cfg(feature = "kzg-rs")] {
-        pub use kzg_rs::KzgSettings;
+        pub use kzg_rs::{KzgSettings,EnvKzgSettings};
     }
 }
-
-pub use env_settings::EnvKzgSettings;
-pub use trusted_setup_points::{
-    parse_kzg_trusted_setup, G1Points, G2Points, KzgErrors, BYTES_PER_G1_POINT, BYTES_PER_G2_POINT,
-    G1_POINTS, G2_POINTS, NUM_G1_POINTS, NUM_G2_POINTS,
-};
